Add error when saving bots that already exist

+++ src/marvin/bots/base.py
@@ -5,6 +5,7 @@ import re
 from typing import TYPE_CHECKING, Any, Callable
 
 import pendulum
+from fastapi import HTTPException, status
 from openai.error import InvalidRequestError
 from pydantic import Field, validator
 
@@ -211,6 +212,7 @@ class Bot(MarvinBaseModel, LoggerMixin):
     @classmethod
     def from_bot_config(cls, bot_config: "BotConfig") -> "Bot":
         return cls(
+            id=bot_config.id,
             name=bot_config.name,
             personality=bot_config.personality,
             instructions=bot_config.instructions,
@@ -218,11 +220,28 @@ class Bot(MarvinBaseModel, LoggerMixin):
             input_transformers=bot_config.input_transformers,
         )
 
-    async def save(self):
-        """Save this bot in the database. Overwrites any existing bot with the
-        same name."""
+    async def save(self, overwrite: bool = False):
+        """
+        Save this bot in the database. Bots are saved by name. If
+        overwrite=False, an error is raised if a bot with the same name
+        exists.
+        """
         bot_config = self.to_bot_config()
-        await marvin.api.bots.delete_bot_config(name=self.name)
+        try:
+            existing_bot = await marvin.api.bots.get_bot_config(name=self.name)
+            if existing_bot and not overwrite:
+                raise ValueError(
+                    f"Bot with name {self.name} already exists. `overwrite=True` to"
+                    " overwrite."
+                )
+            else:
+                await marvin.api.bots.delete_bot_config(name=self.name)
+        except HTTPException as exc:
+            if exc.status_code == status.HTTP_404_NOT_FOUND:
+                pass
+            else:
+                raise
+
         await marvin.api.bots.create_bot_config(bot_config=bot_config)
 
     @classmethod
